+++ deno.json
@@ -20,6 +20,8 @@
   },
   "exports": {
     ".": "./mod.ts",
+    "./layout": "./layout.ts",
+    "./input": "./input.ts",
     "./validate": "./validate.ts"
   },
   "nodeModulesDir": "auto",
