+++ input-native.ts
@@ -169,7 +169,7 @@ export interface InputNative {
   delay(st: number): number;
 }
 
-import { compiled } from "./wasm.ts";
+import { compiled } from "./wasm-input.ts";
 
 export async function createInputNative(
   escLatency: number,
@@ -178,14 +178,6 @@ export async function createInputNative(
 
   let instance = await WebAssembly.instantiate(compiled, {
     env: { memory },
-    clay: {
-      measureTextFunction() {},
-      queryScrollOffsetFunction(ret: number) {
-        let v = new DataView(memory.buffer);
-        v.setFloat32(ret, 0, true);
-        v.setFloat32(ret + 4, 0, true);
-      },
-    },
   });
 
   let exports = instance.exports as unknown as {
